+++ spi/spi_master_slave/spi_master/utilities/hardcoded_functions.hpp
@@ -6,7 +6,7 @@
 // WARNING HARDCODED FUNCTIONS
 void setDefaultSettings();
 
-void set_Freq(uint32_t freq); //установка заданной частоты генератора
+void set_Freq(uint16_t freq); //установка заданной частоты генератора
 
 void init_SPI(uint8_t port ,uint8_t v2 ,uint8_t v3, uint8_t v4); //инициирование SPI
